+++ lib/typescript/render/providers/facebook/FacebookRender.tsx
@@ -16,7 +16,7 @@ import {ButtonTemplate} from './components/ButtonTemplate';
 import {GenericTemplate} from './components/GenericTemplate';
 import {MediaTemplate} from './components/MediaTemplate';
 import {FallbackAttachment} from './components/FallbackAttachment';
-import {StoryMention} from './components/StoryMention';
+import {StoryMention} from './components/InstagramStoryMention';
 
 export const FacebookRender = (props: RenderPropsUnion) => {
   const message = props.message;
@@ -25,7 +25,6 @@ export const FacebookRender = (props: RenderPropsUnion) => {
 };
 
 function render(content: ContentUnion, props: RenderPropsUnion) {
-  console.log('content', content)
   switch (content.type) {
     case 'text':
       return <Text fromContact={props.message.fromContact || false} text={content.text} />;
@@ -65,9 +64,9 @@ function render(content: ContentUnion, props: RenderPropsUnion) {
       );
 
     case 'story_mention':
-      return(
-        <StoryMention  url={content.url} sentAt={content.sentAt} fromContact={props.message.fromContact || false} />
-      )
+      return (
+        <StoryMention url={content.url} sentAt={content.sentAt} fromContact={props.message.fromContact || false} />
+      );
 
     default:
       return null;
@@ -75,11 +74,8 @@ function render(content: ContentUnion, props: RenderPropsUnion) {
 }
 
 const parseAttachment = (
-  attachment: SimpleAttachment | ButtonAttachment | GenericAttachment | MediaAttachment 
+  attachment: SimpleAttachment | ButtonAttachment | GenericAttachment | MediaAttachment
 ): AttachmentUnion => {
-
-  console.log('attachment', attachment)
-
   if (attachment.type === 'image') {
     return {
       type: 'image',
@@ -135,10 +131,6 @@ const parseAttachment = (
 
 function facebookInbound(message): ContentUnion {
   const messageJson = message.content.message ?? message.content;
-  
-console.log('message.content.sentAt', message.sentAt instanceof Date)
-console.log('message', message)
-  console.log('inboud', messageJson)
 
   if (messageJson.attachment?.type === 'fallback' || messageJson.attachments?.[0].type === 'fallback') {
     return {
@@ -160,7 +152,7 @@ console.log('message', message)
     return {
       type: 'story_mention',
       url: messageJson.attachment?.payload?.url || messageJson.attachments[0]?.payload?.url || null,
-     sentAt: message.sentAt
+      sentAt: message.sentAt,
     };
   }
 
@@ -192,8 +184,6 @@ console.log('message', message)
 function facebookOutbound(message): ContentUnion {
   const messageJson = message.content.message ?? message.content;
 
-  console.log('outboud', messageJson)
-
   if (messageJson.quick_replies) {
     if (messageJson.quick_replies.length > 13) {
       messageJson.quick_replies = messageJson.quick_replies.slice(0, 13);
@@ -218,10 +208,9 @@ function facebookOutbound(message): ContentUnion {
     return {
       type: 'story_mention',
       url: messageJson.attachment.url || messageJson.attachments[0].url,
-     sentAt: messageJson.sentAt
+      sentAt: messageJson.sentAt,
     };
   }
-
 
   if (messageJson.attachment?.type === 'fallback' || messageJson.attachments?.[0].type === 'fallback') {
     return {
@@ -248,7 +237,6 @@ function facebookOutbound(message): ContentUnion {
       text: messageJson.text,
     };
   }
-
 
   return {
     type: 'text',
